Creacion de un nuevo main vacio para poder usarlo de forma generica.
