+++ LearningCpp/LearningCpp.vcxproj.filters
@@ -23,6 +23,9 @@
     <ClCompile Include="Point2D.cpp">
       <Filter>Archivos de origen</Filter>
     </ClCompile>
+    <ClCompile Include="CheckingVersion.cpp">
+      <Filter>Archivos de origen</Filter>
+    </ClCompile>
     <ClCompile Include="main.cpp">
       <Filter>Archivos de origen</Filter>
     </ClCompile>
